pyproject.toml
[project]
name = "LiveEventMonitor"
version = "0.1.0"
description = "Tracks Progress of WhatTrivia Events"
authors = [{name = "bernacki", email = "bernackimark@gmail.com"}]
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "Authlib>=1.5.2",
    "google-api-python-client == 2.166.0",
    "gspread>=6.2.0",
    "oauth2client>=4.1.3",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "pytz>=2025.2",
    "streamlit>=1.44.1",
]
